– Overview –
Working on the inventory implementation in the game, as well the document restructurisation.

– Inventory –

Deleted the original file and assigned the room methods in Rooms.cs class, the Player assignments are added to the Player.cs.

– Story –

Story file is restructured in order to create the more abstract and encapsulated version of software, so only the required features are shown public, where the maths and private assignments should be kept private.

This approach ensures that clarity as well as modularity of the solution provided.

– To-Dos –

As you may notice earlier, I have used Rider's built-in TODO highlight feature. I have kept certain tasks active, since I want to update my progress in small increments rather than a one massive solution, where a lot of changes should be tracked and controlled. I believe in the approach of active progress tracking in order to create the best possible solution.