Changing issues Myk has highlighted in his review, such as:
1.First thing I've noticed when running the code is that everytime the code runs the "ENEMY IS INJURED!" (on a fresh start) OR "S HIT HAS MISSED" (after starting again when dying to an enemy) message is displayed.

Issue was with the test file, which contradicted the story file, hence needed changing.

2. Winning a fight against an enemy without a sword also seems impossible so maybe try to balance the game out a little as unless the player manages to get a weapon they have a very small chance of actually winning.

Changed the initial value to 15, so when user finds a sword, it's a small increment, but it's significant enought to feel the difference.

3. If you die your health can go into negatives and it shows up as a negative after the enemy turn so maybe do some validation on player health e.g if playerHealth < 0 then playerHealth = 0.

I changed this with a validation files, that switches your health to 0 and kills you, if that's the case,

4. Finding a sword,axe or biggleswade potion decreases the player's health by the damage value of the item. Image below: (My health was 15 after fighting an enemy.)

Fixed this by creating a simple check for the player's health and not allowing this kind of thing happening in code.

5. For some reason the game also puts me into an exit search multiple times in a row with each time resulting in a no exit found, I think that you should change it so that an exit search can only happen once per room so that the player doen't get stuck in an endless loop.

Changed the randomisation seed to a less unique one, so this shouldn't be the case.